Returns buffer instead of WPARAM and LPARAM

+++ atom/browser/native_window.cc
@@ -469,8 +469,8 @@ void NativeWindow::NotifyWindowMessage(UINT message, WPARAM w_param,
                                        LPARAM l_param) {
   FOR_EACH_OBSERVER(
       NativeWindowObserver, observers_,
-      OnWindowMessage(message, static_cast<uint64_t>(w_param),
-                      static_cast<int64_t>(l_param)));
+      OnWindowMessage(message, w_param,
+                      l_param));
 }
 #endif
